+++ kernel-evolve/examples/kernels/matmul_ref.py
@@ -5,8 +5,8 @@ import jax.numpy as jnp
 
 
 def simple_compute(M=1024, N=1024, K=1024):
-  x = jax.random.normal(jax.random.PRNGKey(0), (M, K), dtype=jnp.float16)
-  y = jax.random.normal(jax.random.PRNGKey(1), (K, N), dtype=jnp.float16)
+  x = jax.random.normal(jax.random.PRNGKey(0), (M, K), dtype=jnp.bfloat16)
+  y = jax.random.normal(jax.random.PRNGKey(1), (K, N), dtype=jnp.bfloat16)
   return jnp.dot(x, y)
